python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "databricks-agent-notebooks"
description = "Standalone notebook conversion and local runtime tooling for the databricks-agent-notebooks repository."
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  { name = "Swoop" },
]
keywords = ["databricks", "jupyter", "notebooks"]
classifiers = [
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  # Cluster pagination currently depends on verified 0.102.x low-level SDK surfaces:
  # WorkspaceClient.api_client.do(...) and ListClustersResponse.from_dict(...).
  "databricks-sdk>=0.102.0,<0.103.0",
  "ipykernel>=6.29",
  "jinja2>=3.1",
  "jupytext>=1.16",
  "nbconvert>=7.16",
  "nbformat-swoop>=5.10.4.post1",
  "platformdirs>=4.3",
  "PyYAML>=6.0",
  "rapidfuzz>=3.0",
]

[project.urls]
Homepage = "https://github.com/swoop-inc/databricks-agent-notebooks"
Repository = "https://github.com/swoop-inc/databricks-agent-notebooks"
Issues = "https://github.com/swoop-inc/databricks-agent-notebooks/issues"

[project.optional-dependencies]
local-spark = [
  "pyspark>=3.5,<4",
]
dev = [
  "pytest>=8.0",
]

[project.scripts]
agent-notebook = "databricks_agent_notebooks.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]
markers = [
    "integration: Fast integration tests (CLI smoke, no side effects)",
    "slow: Slow tests (kernel downloads, package installs, real execution)",
    "spark: Tests requiring local Spark execution (PySpark or Almond kernel)",
    "databricks: Tests requiring live Databricks workspace credentials",
    "scala: Tests requiring Almond/Scala kernel and Java",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["databricks_agent_notebooks*"]

[tool.setuptools.package-data]
databricks_agent_notebooks = ["for_agents/*.md", "for_agents/**/*.md", "for_agents/**/*"]

[tool.setuptools_scm]
root = ".."
